% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{earthquake}
\alias{earthquake}
\title{NOOA 's earthquake data set}
\format{A data frame with large number of columns}
\source{
\url{https://www.ngdc.noaa.gov/nndc/struts/form?t=101650&s=1&d=1}
}
\usage{
earthquake()
}
\description{
The Significant Earthquake Database contains information on destructive earthquakes from 2150 B.C. to the present that meet at least one of the following criteria: Moderate damage (approximately $1 million or more), 10 or more deaths, Magnitude 7.5 or greater, Modified Mercalli Intensity X or greater, or the earthquake generated a tsunami.
}
\examples{
\dontrun{
earthquake
}
}
