% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{eq_clean_date}
\alias{eq_clean_date}
\title{Clean dates from earthquak data}
\usage{
eq_clean_date(data)
}
\arguments{
\item{data}{A data.frame.}
}
\value{
The input data.frame plus a date column.
}
\description{
Clean dates from earthquak data
}
\examples{
\dontrun{
fp <- system.file("exdata", "earthquake.csv", package = "capsmsdr")
da <- read_delim(fp, delim = "\\t") \%>\%
  eq_clean_date \%>\%
}
}
