% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{eq_clean_location}
\alias{eq_clean_location}
\title{Clean locations from earthquak data}
\usage{
eq_clean_location(data)
}
\arguments{
\item{data}{A data.frame.}
}
\value{
The input data.fram plus a location column
}
\description{
This function use some fnctions from stringr package to fix data fom location names in original data.frame.
}
\examples{
\dontrun{
file_path <- system.file("data", "earthquake.csv", package = "capsmsdr")
signif <- read_delim(file_path, delim = "\\t") \%>\%
  eq_clean_date \%>\%
  eq_clean_location
}
}
