% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet.R
\name{eq_create_label}
\alias{eq_create_label}
\title{Create labels for earthquak data}
\usage{
eq_create_label(df)
}
\arguments{
\item{df}{A data.frame.}
}
\value{
A column of earthquak annotation; locaion, magnitude and deaths. Removes NA.
}
\description{
Create labels for earthquak data
}
\examples{
\dontrun{
# Read and clean data

fp <- system.file("data", "earthquake.csv", package = "capsmsdr")

signif <- read_tsv(fp) \%>\%
  eq_clean_date \%>\%
  eq_clean_location

# plot

signif \%>\%
filter(COUNTRY == 'MEXICO' & year(date) >= 2000) \%>\%
  mutate(popup_text = eq_create_label(.)) \%>\%
  eq_map()
}
}
