% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet.R
\name{eq_map}
\alias{eq_map}
\title{Map earthquak data}
\usage{
eq_map(data)
}
\arguments{
\item{data}{A data.frame}
}
\value{
A plot with info from location, magnitude and deaths.
}
\description{
Map earthquak data
}
\examples{
\dontrun{
# Read and clean data
fp <- system.file("exdata", "earthquake.csv", package = "capsmsdr")
signif <- read_tsv(fp) \%>\%
  eq_clean_date \%>\%
  eq_clean_location

# Plot

signif \%>\%
filter(COUNTRY == 'MEXICO' & year(date) >= 2000) \%>\%
  mutate(popup_text = eq_create_label(.)) \%>\%
  eq_map()
}
}
