% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeline.R
\name{geom_timeline}
\alias{geom_timeline}
\title{Earthquak graph with timeline}
\usage{
geom_timeline(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, ...)
}
\arguments{
\item{mapping}{aethetics by aes}

\item{data}{A data.fram}

\item{stat}{A name from a statistical transfromation}

\item{position}{Position adjust}

\item{na.rm}{Remove missing values}

\item{show.legend}{Show legend, TRUE or FALSE}

\item{inherit.aes}{Override default aesthetics, TRUE or FALSE}

\item{...}{Extra arguments}
}
\description{
Earthquak graph with timeline
}
\details{
A timeline plot earthquaks by dates. Colors represents number of deaths and size is the magnitude of the earthquaks.
}
\examples{
\dontrun{
# Read and clean data

fp <- system.file("exdata", "earthquake.csv", package = "capsmsdr")

signif <- read_tsv(fp) \%>\%
  eq_clean_date \%>\%
  eq_clean_location

# Plot
signif \%>\%
filter(COUNTRY == 'MEXICO') \%>\%
  ggplot(aes(date = date,
             xmin = as.Date('1995-01-01'),
             xmax = as.Date('2000-12-30'),
             y = COUNTRY,
             colour = DEATHS,
             fill = DEATHS,
             size = EQ_PRIMARY,
             location = LOCATION)) +
  geom_timeline() +
  geom_timeline_label() +
  theme(axis.line.y = element_blank(),
        axis.line.x = element_line(),
        axis.ticks.y = element_blank(),
        axis.text.y = element_blank(),
        axis.title.y = element_blank(),
        legend.position = 'bottom',
        panel.grid = element_blank(),
        panel.background = element_blank())
}
}
