% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeline.R
\docType{data}
\name{GeomTimeline}
\alias{GeomTimeline}
\title{Geom configuration for ploting timeline}
\format{An object of class \code{GeomTimeline} (inherits from \code{Geom}, \code{ggproto}) of length 5.}
\usage{
GeomTimeline
}
\description{
Geom configuration for ploting timeline
}
\keyword{datasets}
