% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeline.R
\docType{data}
\name{GeomTimelineLabel}
\alias{GeomTimelineLabel}
\title{Geom for timeline labels}
\format{An object of class \code{GeomTimeline} (inherits from \code{Geom}, \code{ggproto}) of length 3.}
\usage{
GeomTimelineLabel
}
\description{
Geom for timeline labels
}
\keyword{datasets}
